app-group-reporting/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
databricks-sdk>=0.81.0
pydantic>=2.0.0
